+++ CHANGELOG.md
@@ -8,6 +8,7 @@ with `zswap` being tracked in [Changelog Zswap](./CHANGELOG_zswap.md).
 
 - Remove special-casing of validation behaviour depending on the
   `test-utilities` feature being present.
+- Change ledger `DustSpendError::BackingNightNotFound` enum variant to now hold its data value on the heap (to reduce Enum size), i.e. this variant is now defined as: `BackingNightNotFound(Box<QualifiedDustOutput>)`.
 
 ## 6.1.0